backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
pylovepdf==1.3.3
gunicorn==21.2.0
python-dotenv==1.0.0